requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
backcall==0.1.0
behave==1.2.6
cymem==1.31.2
cytoolz==0.8.2
de-core-news-sm==2.0.0
decorator==4.3.0
dill==0.2.7.1
funcy==1.10.2
future==0.16.0
ipython==6.3.1
ipython-genutils==0.2.0
jedi==0.12.0
Jinja2==2.10
joblib==0.11
MarkupSafe==1.0
more-itertools==4.1.0
msgpack-numpy==0.4.1
msgpack-python==0.5.6
murmurhash==0.28.0
numexpr==2.6.5
numpy==1.14.3
pandas==0.22.0
parse==1.8.4
parse-type==0.4.2
parso==0.2.0
pathlib==1.0.1
pexpect==4.5.0
pickleshare==0.7.4
plac==0.9.6
pluggy==0.6.0
preshed==1.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.5.3
Pygments==2.2.0
PyHamcrest==1.9.0
pyLDAvis==2.1.1
pytest==3.5.1
python-dateutil==2.7.3
pytz==2018.4
regex==2017.4.5
scikit-learn==0.19.1
scipy==1.1.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
spacy==2.0.11
termcolor==1.1.0
thinc==6.10.2
toolz==0.9.0
tqdm==4.23.2
traitlets==4.3.2
ujson==1.35
wcwidth==0.1.7
wrapt==1.10.11
